+++ packages/db/src/schema/instructors.ts
@@ -1,0 +1,40 @@
+import { sql } from "drizzle-orm";
+import { index, pgTable, uniqueIndex, uuid, varchar } from "drizzle-orm/pg-core";
+import { websocInstructor } from "./websoc.ts";
+
+export const instructor = pgTable(
+  "instructor",
+  {
+    ucinetid: varchar("ucinetid").primaryKey(),
+    name: varchar("name").notNull(),
+    title: varchar("title").notNull(),
+    email: varchar("email").notNull(),
+    department: varchar("department").notNull(),
+  },
+  (table) => [
+    index("instructor_search_index").using(
+      "gin",
+      sql`(
+ SETWEIGHT(TO_TSVECTOR('english', COALESCE(${table.ucinetid}, '')), 'A') ||
+ SETWEIGHT(TO_TSVECTOR('english', COALESCE(${table.name}, '')), 'B') ||
+ SETWEIGHT(TO_TSVECTOR('english', COALESCE(${table.title}, '')), 'B')
+)`,
+    ),
+  ],
+);
+
+export const instructorToWebsocInstructor = pgTable(
+  "instructor_to_websoc_instructor",
+  {
+    id: uuid("id").primaryKey().defaultRandom(),
+    instructorUcinetid: varchar("instructor_ucinetid").references(() => instructor.ucinetid),
+    websocInstructorName: varchar("websoc_instructor_name")
+      .notNull()
+      .references(() => websocInstructor.name),
+  },
+  (table) => [
+    index().on(table.instructorUcinetid),
+    index().on(table.websocInstructorName),
+    uniqueIndex().on(table.instructorUcinetid, table.websocInstructorName),
+  ],
+);
